+++ rest-errors/src/main/java/com/truward/brikar/error/RestErrorCode.java
@@ -7,6 +7,8 @@ package com.truward.brikar.error;
  */
 public interface RestErrorCode {
 
+  int getHttpStatus();
+
   String getCodeName();
 
   String getDescription();